fix(rules): 🐛  replace ReadFileAsJson with ReadFile
ReadFileAsJson would only be useful if you must use the return value as an object within the same template. In this case, "value" will always turn out to be a string, and by using "type: map", essentially, we have only 1 option, which is unmarshal as JSON. No need to unmarshal twice.